// apps/web/generated/api/models/fileAssetResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FileAssetContext } from './fileAssetContext';

export interface FileAssetResponse {
  file_asset_id: string;
  file_name: string;
  size_bytes: number;
  mime_type: string;
  uploaded_at: string;
  context: FileAssetContext;
}