// apps/web/generated/api/models/resumeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeResponse {
  resume_id: string;
  freelancer_profile_id: string;
  file_asset_id: string;
  version_no: number;
  summary: string | null;
  is_current: boolean;
}