// apps/web/generated/api/models/fileAssetContext.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export type FileAssetContext = typeof FileAssetContext[keyof typeof FileAssetContext];


export const FileAssetContext = {
  resume: 'resume',
  portfolio: 'portfolio',
  delivery: 'delivery',
  ticket_attachment: 'ticket_attachment',
  generic: 'generic',
} as const;